+++ src/__init__.py
@@ -1,2 +1,2 @@
 # OpenCR Server
-__version__ = "0.2.1"
+__version__ = "0.3.0"